Moved absmag_comovingdist_functions.py to legacy folder, changed dist_funcs_cython.pyx to distance.pyx
