+++ VoidFinder/python/voidfinder/table_functions.py
@@ -8,6 +8,10 @@ General astropy table functions.
 import numpy as np
 from astropy.table import Table
 
+
+
+
+
 def add_row(table, row):
     '''Column-wise add table row to table'''
